requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
pillow>=9.0.0
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.65.0
scikit-learn>=1.2.0
accelerate>=0.20.0 